refactor(shared-decks): move into a shareddeck package
