+++ AnkiDroid/src/main/java/com/ichi2/anki/IntentHandler.kt
@@ -32,6 +32,7 @@ import com.ichi2.anki.dialogs.requireDeckPickerOrShowError
 import com.ichi2.anki.exception.SystemStorageException
 import com.ichi2.anki.libanki.DeckId
 import com.ichi2.anki.settings.Prefs
+import com.ichi2.anki.shareddeck.SharedDecksDownloadFragment
 import com.ichi2.anki.ui.windows.reviewer.ReviewerFragment
 import com.ichi2.anki.utils.MimeTypeUtils
 import com.ichi2.anki.worker.SyncWorker
