+++ AnkiDroid/src/test/java/com/ichi2/anki/shareddeck/SharedDecksScreenshotTest.kt
@@ -1,6 +1,6 @@
 // SPDX-License-Identifier: GPL-3.0-or-later
 
-package com.ichi2.anki
+package com.ichi2.anki.shareddeck
 
 import android.app.DownloadManager
 import android.content.Intent
@@ -17,6 +17,8 @@ import androidx.core.view.WindowInsetsCompat.Type.displayCutout
 import androidx.core.view.WindowInsetsCompat.Type.navigationBars
 import androidx.core.view.WindowInsetsCompat.Type.statusBars
 import androidx.fragment.app.commit
+import com.ichi2.anki.R
+import com.ichi2.anki.ScreenshotTest
 import com.ichi2.testutils.insetsOf
 import com.ichi2.utils.dp
 import org.junit.Test
@@ -29,7 +31,7 @@ import org.robolectric.RuntimeEnvironment
 /**
  * Screenshot tests for [SharedDecksActivity]
  *
- * `./gradlew :AnkiDroid:verifyRoborazziPlayDebug -Pscreenshot --tests "com.ichi2.anki.SharedDecksScreenshotTest"`
+ * `./gradlew :AnkiDroid:verifyRoborazziPlayDebug -Pscreenshot --tests "com.ichi2.anki.shareddeck.SharedDecksScreenshotTest"`
  */
 class SharedDecksScreenshotTest : ScreenshotTest() {
     @Test
